refactor: move test code
