+++ embrace-android-core/src/test/java/io/embrace/android/embracesdk/internal/injection/DependencyInjectionKtTest.kt
@@ -1,8 +1,5 @@
-package io.embrace.android.embracesdk.injection
+package io.embrace.android.embracesdk.internal.injection
 
-import io.embrace.android.embracesdk.internal.injection.LoadType
-import io.embrace.android.embracesdk.internal.injection.factory
-import io.embrace.android.embracesdk.internal.injection.singleton
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.util.concurrent.atomic.AtomicInteger
