+++ embrace-android-core/src/test/java/io/embrace/android/embracesdk/internal/injection/MomentsModuleImplTest.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.injection
+package io.embrace.android.embracesdk.internal.injection
 
 import io.embrace.android.embracesdk.fakes.FakeConfigModule
 import io.embrace.android.embracesdk.fakes.injection.FakeDeliveryModule
@@ -6,7 +6,6 @@ import io.embrace.android.embracesdk.fakes.injection.FakeEssentialServiceModule
 import io.embrace.android.embracesdk.fakes.injection.FakeInitModule
 import io.embrace.android.embracesdk.fakes.injection.FakePayloadSourceModule
 import io.embrace.android.embracesdk.fakes.injection.FakeWorkerThreadModule
-import io.embrace.android.embracesdk.internal.injection.createMomentsModule
 import org.junit.Assert.assertNotNull
 import org.junit.Test
