+++ embrace-android-core/src/test/java/io/embrace/android/embracesdk/internal/session/message/PayloadMessageCollatorImplTest.kt
@@ -1,13 +1,15 @@
 package io.embrace.android.embracesdk.internal.session.message
 
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.embrace.android.embracesdk.fakes.FakeEnvelopeMetadataSource
 import io.embrace.android.embracesdk.fakes.FakeEnvelopeResourceSource
 import io.embrace.android.embracesdk.fakes.FakeGatingService
 import io.embrace.android.embracesdk.fakes.FakePreferenceService
 import io.embrace.android.embracesdk.fakes.FakeSessionPayloadSource
-import io.embrace.android.embracesdk.fakes.injection.FakeCoreModule
 import io.embrace.android.embracesdk.fakes.injection.FakeInitModule
 import io.embrace.android.embracesdk.internal.envelope.session.SessionEnvelopeSourceImpl
+import io.embrace.android.embracesdk.internal.injection.CoreModule
+import io.embrace.android.embracesdk.internal.injection.CoreModuleImpl
 import io.embrace.android.embracesdk.internal.payload.ApplicationState
 import io.embrace.android.embracesdk.internal.payload.Envelope
 import io.embrace.android.embracesdk.internal.payload.LifeEventType
@@ -20,11 +22,14 @@ import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RuntimeEnvironment
 
+@RunWith(AndroidJUnit4::class)
 internal class PayloadMessageCollatorImplTest {
 
     private lateinit var initModule: FakeInitModule
-    private lateinit var coreModule: FakeCoreModule
+    private lateinit var coreModule: CoreModule
     private lateinit var currentSessionSpan: CurrentSessionSpan
     private lateinit var collator: PayloadMessageCollatorImpl
     private lateinit var gatingService: FakeGatingService
@@ -32,7 +37,7 @@ internal class PayloadMessageCollatorImplTest {
     @Before
     fun setUp() {
         initModule = FakeInitModule()
-        coreModule = FakeCoreModule()
+        coreModule = CoreModuleImpl(RuntimeEnvironment.getApplication(), initModule.logger)
         gatingService = FakeGatingService()
         val sessionEnvelopeSource = SessionEnvelopeSourceImpl(
             metadataSource = FakeEnvelopeMetadataSource(),
