+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeNativeFeatureModule.kt
@@ -1,8 +1,5 @@
-package io.embrace.android.embracesdk.fakes.injection
+package io.embrace.android.embracesdk.fakes
 
-import io.embrace.android.embracesdk.fakes.FakeClock
-import io.embrace.android.embracesdk.fakes.FakeNativeCrashService
-import io.embrace.android.embracesdk.fakes.FakeNdkService
 import io.embrace.android.embracesdk.internal.anr.ndk.NativeAnrOtelMapper
 import io.embrace.android.embracesdk.internal.anr.ndk.NativeThreadSamplerInstaller
 import io.embrace.android.embracesdk.internal.anr.ndk.NativeThreadSamplerService
