+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeSessionSpanAttrPopulator.kt
@@ -4,7 +4,7 @@ import io.embrace.android.embracesdk.internal.payload.LifeEventType
 import io.embrace.android.embracesdk.internal.payload.SessionZygote
 import io.embrace.android.embracesdk.internal.session.orchestrator.SessionSpanAttrPopulator
 
-internal class FakeSessionSpanAttrPopulator : SessionSpanAttrPopulator {
+public class FakeSessionSpanAttrPopulator : SessionSpanAttrPopulator {
 
     override fun populateSessionSpanStartAttrs(session: SessionZygote) {
     }
